refactor: Componentization member register page
멤버 등록하기 페이지를 컴포넌트화 하였습니다.
